refactor(engine): rename core module to tracking for clarity
