+++ packages/devqubit-braket/src/devqubit_braket/envelope.py
@@ -57,7 +57,7 @@ from devqubit_engine.utils.serialization import to_jsonable
 
 
 if TYPE_CHECKING:
-    from devqubit_engine.core.run import Run
+    from devqubit_engine.tracking.run import Run
 
 
 logger = logging.getLogger(__name__)
