+++ packages/devqubit-braket/src/devqubit_braket/snapshot.py
@@ -35,7 +35,7 @@ from devqubit_engine.utils.common import utc_now_iso
 
 
 if TYPE_CHECKING:
-    from devqubit_engine.core.run import Run
+    from devqubit_engine.tracking.run import Run
 
 logger = logging.getLogger(__name__)
