+++ packages/devqubit-braket/src/devqubit_braket/tracked.py
@@ -20,7 +20,7 @@ from devqubit_braket.utils import extract_task_id
 
 
 if TYPE_CHECKING:
-    from devqubit_engine.core.run import Run
+    from devqubit_engine.tracking.run import Run
     from devqubit_engine.uec.envelope import ExecutionEnvelope
