+++ packages/devqubit-engine/src/devqubit_engine/storage/backends/local.py
@@ -24,7 +24,6 @@ import tempfile
 from pathlib import Path
 from typing import Any, Iterator
 
-from devqubit_engine.core.record import RunRecord
 from devqubit_engine.storage.types import (
     ArtifactRef,
     BaselineInfo,
@@ -32,6 +31,7 @@ from devqubit_engine.storage.types import (
     RunNotFoundError,
     RunSummary,
 )
+from devqubit_engine.tracking.record import RunRecord
 from devqubit_engine.utils.common import utc_now_iso
