+++ packages/devqubit-engine/src/devqubit_engine/uec/api/synthesize.py
@@ -41,8 +41,8 @@ from devqubit_engine.utils.common import is_manual_run_record
 
 
 if TYPE_CHECKING:
-    from devqubit_engine.core.record import RunRecord
     from devqubit_engine.storage.types import ArtifactRef, ObjectStoreProtocol
+    from devqubit_engine.tracking.record import RunRecord
 
 
 logger = logging.getLogger(__name__)
