+++ packages/devqubit-pennylane/src/devqubit_pennylane/adapter.py
@@ -51,7 +51,7 @@ import types
 import uuid
 from typing import Any
 
-from devqubit_engine.core.run import Run
+from devqubit_engine.tracking.run import Run
 from devqubit_engine.uec.device import DeviceSnapshot
 from devqubit_engine.uec.envelope import ExecutionEnvelope
 from devqubit_engine.uec.execution import ExecutionSnapshot
