+++ tests/conftest.py
@@ -19,9 +19,9 @@ import pytest
 from click.testing import CliRunner
 from devqubit_engine.cli import cli
 from devqubit_engine.config import Config
-from devqubit_engine.core.record import RunRecord
 from devqubit_engine.storage.backends.local import LocalRegistry, LocalStore
 from devqubit_engine.storage.types import ArtifactRef
+from devqubit_engine.tracking.record import RunRecord
 from devqubit_engine.utils.common import utc_now_iso
