+++ tests/test_cli.py
@@ -13,8 +13,8 @@ import json
 from pathlib import Path
 from typing import Any, Callable
 
-from devqubit_engine.core.record import RunRecord
 from devqubit_engine.storage.backends.local import LocalRegistry
+from devqubit_engine.tracking.record import RunRecord
 
 
 # =============================================================================
